Added reacts / liking
Added the ability to react / like to a post. This is built to be purely optional and will not affect groups who have not / don't want to implement this feature. 

Also it can be a bit complicated / convoluted in places. Especially in the update to post content. The reason it is formatted in that way is to make it as easy as 
possible to implement basic features like listing the amount of each react on a post.